Issue #32 - rework display scale / font atlas updating
Prior to this change, the font atlas was initialised once during startup and then display scale changes were handled by the CPU thread at the beginning of the frame. But despite display scale having an impact on font configuration, the font was not updated to reflect new display scales.

This issue came about in Bevy 0.14.0, where it is no longer possible to know the display scale up front during intialisation.

Now, all display scale related context data (including the font data) is initialised once during initialisation, and then updated during the 'extract' stage. This is the only safe point at which we can update the font and request the font atlas to be recalculated, since we know the renderer hasn't began rendering the next frame at this point.