Issue #39 - imgui_new_frame_system should always start a new ImGui frame
Prior to this change, we would exit early if the primary window was not available. In Bevy 0.14.1, the system update order in the Update phase was updated such that a new frame is now fully evaluated having destroyed the primary window. Because we were exiting early, we weren't starting a new ImGui frame in this case, and this lead to other systems triggering an ImGui assert if they attempt to interact with the ImGui context. To guard against this, we now always return from imgui_new_frame_system having began a new ImGui frame.